[Feat] 상단 네비게이션 바 추가 (#19)
