+++ DATEROAD-iOS/DATEROAD-iOS/Presentation/CourseDetail/Models/CourseDetailContents.swift
@@ -50,13 +50,15 @@ extension CourseDetailContents {
         UIImage(resource: .image1)
     ]
     
-    static let timelineContents: [CourseDetailContents] = [
-        CourseDetailContents(index: 1, location: "성수 미술관 성수점", time: "1시간"),
-        CourseDetailContents(index: 2, location: "한강 공원", time: "2시간"),
-        CourseDetailContents(index: 3, location: "광화문 광장", time: "1시간"),
-        CourseDetailContents(index: 4, location: "이태원 맛집", time: "2시간"),
-        CourseDetailContents(index: 5, location: "서울 타워", time: "1시간")
-    ]
+    static func timelineContents() -> [CourseDetailContents] {
+        return [
+            CourseDetailContents(index: 1, location: "성수 미술관 성수점", time: "1시간"),
+            CourseDetailContents(index: 2, location: "한강 공원", time: "2시간"),
+            CourseDetailContents(index: 3, location: "광화문 광장", time: "1시간"),
+            CourseDetailContents(index: 4, location: "이태원 맛집", time: "2시간"),
+            CourseDetailContents(index: 5, location: "서울 타워", time: "1시간")
+        ]
+    }
     
     static let hashTagContents: [CourseDetailContents] = [
         CourseDetailContents(hashTag: "🚙 드라이브"),
