+++ DATEROAD-iOS/DATEROAD-iOS/Presentation/CourseDetail/ViewModels/TimelineInfoViewModel.swift
@@ -9,6 +9,6 @@ import Foundation
 
 struct TimelineInfoViewModel {
     var numberOfItems: Int {
-        return 5
+        return 1
     }
 }
